requirements.txt
typer
requests
fake_header
